Fixes for MSVC warnings, less overall changes
Kept some fixes from the first round of review. Found out that several
targets weren't being built with the same basic warnings disabled.
Disabled a few warnings for external libraries specifically, rather than
applying them to all targets.

+++ apps/openmw/mwgui/inventorywindow.hpp
@@ -52,7 +52,7 @@ namespace MWGui
             DragAndDrop* mDragAndDrop;
 
             bool mPreviewDirty;
-            size_t mSelectedItem;
+            int mSelectedItem;
 
             MWWorld::Ptr mPtr;
